refactor: engineLoop.ts
